+++ R/get_project.R
@@ -7,6 +7,7 @@
 #' @param case Character; One of  of the following:
 #' \tabular{llll}{
 #'   \strong{case}  \tab \strong{Description}\tab  \strong{EF} \tab \strong{Notes}   \cr
+#'   argentina  \tab top down\tab COPERT \tab  .rds\cr
 #'   emislacovid  \tab Bottom-up March 2020\tab CETESB \tab  .rds\cr
 #'   brazil_bu_chem     \tab Bottom-up  chemical mechanisms\tab CETESB+tunnel\tab  .rds\cr
 #'   brazil_bu_chem_streets \tab Bottom-up  chemical mechanisms for streets and MUNICH\tab CETESB+tunnel\tab  .rds\cr
@@ -17,8 +18,7 @@
 #'   sebr_cb05co2   \tab Top-down SP, MG and RJ\tab CETESB+tunnel\tab  .rds\cr
 #'   amazon2014     \tab Top-down Amazon\tab CETESB+tunnel\tab  csv and.rds\cr
 #'   curitiba       \tab Bottom-down +GTFS\tab CETESB+tunnel\tab  csv and.rds\cr
-#'   ecuador_td     \tab Top-down. Renamed ecuador_td_im\tab EEA\tab  csv and.rds\cr
-#'   ecuador_td_hot_month  \tab Top-down\tab EEA\tab  csv and.rds\cr
+#'   ecuador     \tab Top-down. Renamed ecuador_td_im\tab EEA\tab  csv and.rds\cr
 #'   moves_bu         \tab Bottom-up\tab US/EPA MOVES \tab  csv and.rds (requires MOVES >=3.0 on Windows)\cr
 #'   manizales_bu     \tab Bottom-up  chemical mechanisms\tab EEA\tab  csv, csv.gz, .rds\cr
 #'   eu_bu_chem       \tab Bottom-up  chemical mechanisms\tab EEA 2019\tab  .rds\cr
@@ -28,37 +28,7 @@
 #' }
 #' @param url String, with the URL to download VEIN project
 #' @note  All projects include option to apply survival functions
-#' @note  \strong{brazil_bu_chem} covers
-#' "brazil",
-#' "brazil_bu",
-#' "brasil_bu",
-#' "brazil_bu_chem",
-#' "brazil_bu_csvgz",
-#' "brazil_bu_csv",
-#' "brazil_bu_cb05",
-#' "brazil_mech",
-#' "brazil_bu_chem_month",
-#' "brazil_bu_chem_im"
-#' "brazil_bu_chem_streets_im" (type <- 'streets')
-#' "brazil_bu_chem_streets" (type <- 'streets')
-#'
-#' \strong{brazil_td_chem} covers
-#' "brazil_td_chem_im"
-#'
-#' \strong{sebr_cb05co2} covers
-#' "sebr_cb05co2_im"
-#'
-#' \strong{ecuador_td} covers
-#' "ecuador_td",
-#' "ecuador_td_hot",
-#' "ecuador_td_im"
-#'
-#' In any case, if you find any error, please, send a pull request in github.
-#'
 #' In Sao Paulo the IM programs was functioning until 2011.
-#'
-#' brazil_countryv2 has scripts updated
-#'
 #' @importFrom utils download.file untar
 #' @export
 #' @examples \dontrun{
@@ -141,6 +111,24 @@ get_project <- function(directory,
       utils::untar(tarfile = tf, exdir = directory)
       message("Your directory is in ", directory)
 
+      # argentina ####
+    } else if(case %in% c("argentina")){
+      URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/argentina.tar.gz"
+      tf <- paste0(tempfile(), ".tar.gz")
+      utils::download.file(url = URL,
+                           destfile =  tf)
+      utils::untar(tarfile = tf, exdir = directory)
+
+      URL <- "https://gitlab.com/ibarraespinosa/veinextras/-/raw/master/wrf_argentina/wrfinput_d01"
+      tf <- paste0(tempfile(), ".tar.gz")
+      utils::download.file(url = URL,
+                           destfile =  paste0(directory, "/wrfinput_d01"))
+
+      message("Your directory is in ", directory)
+
+
+      message("Your directory is in ", directory)
+
       # brazil_bu_chem_streets ####
     } else if(case %in% c("brazil_bu_chem_streets",
                           "brazil_bu_chem_streets_im")){
@@ -178,7 +166,8 @@ get_project <- function(directory,
       message("Your directory is in ", directory)
 
       # ecuador_td_im ####
-    } else if(case %in% c("ecuador_td",
+    } else if(case %in% c("ecuador",
+                          "ecuador_td",
                           "ecuador_td_hot",
                           "ecuador_td_im")){
       URL <- "https://raw.githubusercontent.com/atmoschem/vein/master/projects/ecuador_td.tar.gz"
@@ -187,11 +176,11 @@ get_project <- function(directory,
                            destfile =  tf)
       utils::untar(tarfile = tf, exdir = directory)
 
-      x <- readLines(paste0(directory, "/main.R"))
-      x <- gsub(pattern = "IM <- FALSE",
-                replacement = "IM <- TRUE",
-                x = x)
-      writeLines(x, paste0(directory, "/main.R"))
+      # x <- readLines(paste0(directory, "/main.R"))
+      # x <- gsub(pattern = "IM <- FALSE",
+      #           replacement = "IM <- TRUE",
+      #           x = x)
+      # writeLines(x, paste0(directory, "/main.R"))
 
       message("Your directory is in ", directory)